+++ Mods ChoGGi/Biorobot Specialisation Uniforms/metadata.lua
@@ -1,5 +1,5 @@
 return PlaceObj("ModDef", {
-	"title", "Android Spec Uniforms",
+	"title", "Biorobot Specialisation Uniforms",
 	"id", "ChoGGi_AndroidSpecUniforms",
 	"steam_id", "1780312214",
 	"pops_any_uuid", "577f24f2-5e38-4f5d-8d2c-6e669a2c4f27",
@@ -13,7 +13,7 @@ return PlaceObj("ModDef", {
 		"Code/Script.lua",
 	},
 	"TagCosmetics", true,
-	"description", [[Androids/Biorobots use the correct uniform for their specialisation.
+	"description", [[Biorobots/Androids use the correct uniform for their specialisation.
 
 
 Requested by BLade.]],