+++ Mods ChoGGi/Colonists Force To New Dome/items.lua
@@ -1,0 +1,8 @@
+return {
+	PlaceObj("ModItemOptionToggle", {
+		"name", "CleanupColonists",
+		"DisplayName", T(302535920011767, "Cleanup Colonists"),
+		"Help", T(302535920011768, "Remove any invalid colonists when move forcing."),
+		"DefaultValue", true,
+	}),
+}