+++ Mods ChoGGi/Dome Show Connected Objects/items.lua
@@ -6,4 +6,10 @@ return {
 		"DisplayName", T(302535920011303, "Enable Mod"),
 		"DefaultValue", true,
 	}),
+	PlaceObj("ModItemOptionToggle", {
+		"name", "CleanUpInvalid",
+		"DisplayName", T(302535920011769, "Clean Up Invalid"),
+		"Help", T(302535920011770, "Remove any invalid objects stuck in the dome when you press the toggle connected objs button."),
+		"DefaultValue", false,
+	}),
 }